Add default_scopes decorator
