+++ monitoring/prober/conftest.py
@@ -1,152 +1,9 @@
 import copy
-import requests
-import urllib.parse
 import uuid
-from typing import Dict, List
 
-from google.auth.transport import requests as google_requests
-from google.oauth2 import service_account
+from . import infrastructure
+
 import pytest
-
-ALL_SCOPES = [
-    'dss.write.identification_service_areas',
-    'dss.read.identification_service_areas',
-    'utm.strategic_coordination',
-    'utm.constraint_management',
-    'utm.constraint_consumption'
-]
-
-
-class AuthAdapter(object):
-  """Base class for an adapter that add JWTs to requests."""
-
-  def __init__(self):
-    self._tokens = {}
-
-  def issue_token(self, intended_audience: str, scopes: List[str]) -> str:
-    """Subclasses must return a bearer token for the given audience."""
-
-    raise NotImplementedError()
-
-  def get_headers(self, url: str, scopes: List[str] = None) -> Dict[str, str]:
-    if scopes is None:
-      scopes = ALL_SCOPES
-    intended_audience = urllib.parse.urlparse(url).hostname
-    scope_string = ' '.join(scopes)
-    if intended_audience not in self._tokens:
-      self._tokens[intended_audience] = {}
-    if scope_string not in self._tokens[intended_audience]:
-      self._tokens[intended_audience][scope_string] = self.issue_token(intended_audience, scopes)
-    token = self._tokens[intended_audience][scope_string]
-    return {'Authorization': 'Bearer ' + token}
-
-  def add_headers(self, request: requests.PreparedRequest, scopes: List[str]):
-    for k, v in self.get_headers(request.url, scopes).items():
-      request.headers[k] = v
-
-
-class DummyOAuthServerAdapter(AuthAdapter):
-  """Auth adapter that gets JWTs that uses the Dummy OAuth Server"""
-
-  def __init__(self, token_endpoint: str, sub: str):
-    super().__init__()
-
-    self._oauth_token_endpoint = token_endpoint
-    self._sub = sub
-    self._oauth_session = requests.Session()
-
-  # Overrides method in AuthAdapter
-  def issue_token(self, intended_audience: str, scopes: List[str]) -> str:
-    url = '{}?grant_type=client_credentials&scope={}&intended_audience={}&issuer=dummy&sub={}'.format(
-        self._oauth_token_endpoint, urllib.parse.quote(' '.join(scopes)),
-        urllib.parse.quote(intended_audience), self._sub)
-    response = self._oauth_session.post(url).json()
-    return response['access_token']
-
-class ServiceAccountAuthAdapter(AuthAdapter):
-  """Auth adapter that gets JWTs using a service account."""
-
-  def __init__(self, token_endpoint, service_account_json):
-    super().__init__()
-
-    credentials = service_account.Credentials.from_service_account_file(
-        service_account_json).with_scopes(['email'])
-    oauth_session = google_requests.AuthorizedSession(credentials)
-
-    self._oauth_token_endpoint = token_endpoint
-    self._oauth_session = oauth_session
-
-  # Overrides method in AuthAdapter
-  def issue_token(self, intended_audience: str, scopes: List[str]) -> str:
-    url = '{}?grant_type=client_credentials&scope={}&intended_audience={}'.format(
-        self._oauth_token_endpoint, urllib.parse.quote(' '.join(scopes)),
-        urllib.parse.quote(intended_audience))
-    response = self._oauth_session.post(url).json()
-    return response['access_token']
-
-
-class UsernamePasswordAuthAdapter(AuthAdapter):
-  """Auth adapter that gets JWTs using a username and password."""
-
-  def __init__(self, token_endpoint, username, password, client_id):
-    super().__init__()
-
-    self._oauth_token_endpoint = token_endpoint
-    self._username = username
-    self._password = password
-    self._client_id = client_id
-
-  # Overrides method in AuthAdapter
-  def issue_token(self, intended_audience: str, scopes: List[str]) -> str:
-    scopes.append('aud:{}'.format(intended_audience))
-    response = requests.post(self._oauth_token_endpoint, data={
-      'grant_type': "password",
-      'username': self._username,
-      'password': self._password,
-      'client_id': self._client_id,
-      'scope': ' '.join(scopes),
-    }).json()
-    return response['access_token']
-
-
-class DSSTestSession(requests.Session):
-  """
-  Requests session that provides additional functionality for DSS tests:
-    * Adds a prefix to URLs that start with a '/'.
-    * Automatically applies authorization according to adapter, when present
-  """
-
-  def __init__(self, prefix_url: str, auth_adapter: AuthAdapter = None):
-    super().__init__()
-
-    self._prefix_url = prefix_url
-    self._auth_adapter = auth_adapter
-
-  # Overrides method on requests.Session
-  def prepare_request(self, request, **kwargs):
-    # Automatically prefix any unprefixed URLs
-    if request.url.startswith('/'):
-      request.url = self._prefix_url + request.url
-
-    return super().prepare_request(request, **kwargs)
-
-  def request(self, method, url, **kwargs):
-    if 'auth' not in kwargs and self._auth_adapter:
-      scopes = None
-      if 'scopes' in kwargs:
-        scopes = kwargs['scopes']
-        del kwargs['scopes']
-      if 'scope' in kwargs:
-        scopes = [kwargs['scope']]
-        del kwargs['scope']
-      if scopes is None:
-        scopes = ALL_SCOPES
-      def auth(prepared_request: requests.PreparedRequest) -> requests.PreparedRequest:
-        self._auth_adapter.add_headers(prepared_request, scopes)
-        return prepared_request
-      kwargs['auth'] = auth
-
-    return super().request(method, url, **kwargs)
 
 
 def pytest_addoption(parser):
@@ -176,15 +33,17 @@ def make_auth_adapter(pytestconfig, prefix: str, dummy_oauth_sub: str):
   # Create an auth adapter to get JWTs using the given credentials.  We can use
   # either a service account, a username/password/client_id or a dummy oauth server.
   if pytestconfig.getoption(prefix + '_service_account_json') is not None:
-    auth_adapter = ServiceAccountAuthAdapter(oauth_token_endpoint,
-                                             pytestconfig.getoption(prefix + '_service_account_json'))
+    auth_adapter = infrastructure.ServiceAccountAuthAdapter(
+        oauth_token_endpoint,
+        pytestconfig.getoption(prefix + '_service_account_json'))
   elif pytestconfig.getoption(prefix + '_username') is not None:
-    auth_adapter = UsernamePasswordAuthAdapter(oauth_token_endpoint,
-                                               pytestconfig.getoption(prefix + '_username'),
-                                               pytestconfig.getoption(prefix + '_password'),
-                                               pytestconfig.getoption(prefix + '_client_id'))
+    auth_adapter = infrastructure.UsernamePasswordAuthAdapter(
+        oauth_token_endpoint,
+        pytestconfig.getoption(prefix + '_username'),
+        pytestconfig.getoption(prefix + '_password'),
+        pytestconfig.getoption(prefix + '_client_id'))
   elif pytestconfig.getoption('use_dummy_oauth') is not None:
-    auth_adapter = DummyOAuthServerAdapter(oauth_token_endpoint, dummy_oauth_sub)
+    auth_adapter = infrastructure.DummyOAuthServerAdapter(oauth_token_endpoint, dummy_oauth_sub)
   else:
     raise ValueError(
       'You must provide either an OAuth service account, or a username, '
@@ -201,7 +60,7 @@ def session(pytestconfig):
   api_version_role = pytestconfig.getoption('api_version_role', '')
 
   auth_adapter = make_auth_adapter(pytestconfig, 'oauth', 'fake_uss')
-  s = DSSTestSession(dss_endpoint + api_version_role, auth_adapter)
+  s = infrastructure.DSSTestSession(dss_endpoint + api_version_role, auth_adapter)
   return s
 
 
@@ -212,7 +71,7 @@ def scd_session(pytestconfig):
     return None
 
   auth_adapter = make_auth_adapter(pytestconfig, 'oauth', 'fake_uss')
-  s = DSSTestSession(scd_dss_endpoint, auth_adapter)
+  s = infrastructure.DSSTestSession(scd_dss_endpoint, auth_adapter)
   return s
 
 
@@ -223,7 +82,7 @@ def scd_session2(pytestconfig):
     return None
 
   auth_adapter = make_auth_adapter(pytestconfig, 'oauth', 'fake_uss2')
-  s = DSSTestSession(scd_dss_endpoint, auth_adapter)
+  s = infrastructure.DSSTestSession(scd_dss_endpoint, auth_adapter)
   return s
 
 
@@ -234,7 +93,7 @@ def no_auth_session(pytestconfig):
   if dss_endpoint is None:
     raise ValueError('Missing required --dss-endpoint')
 
-  s = DSSTestSession(dss_endpoint + api_version_role, None)
+  s = infrastructure.DSSTestSession(dss_endpoint + api_version_role, None)
   return s
