+++ monitoring/prober/scd/common.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+import functools
 from typing import Dict, List, Optional, Tuple
